GbDocumentManager: start on document manager to manage live buffers
The long term idea here is that all the open GtkTextBuffer's will be
managed by this. I think we may also add a generic document interface
so that things like glade documents can also be managed here. Not sure
yet though.

Then, we can attach this buffer manager to the combobox in the
GbTabStack. At that point, we can dynamically create GbEditorTab or
GbEditorFrame instances to the documents as needed. You can then just
display any document easily on any tab stack.