gb-source-vim: Add support for vim scrolloff
Vim scrolloff allow the user to set a margin of a number of lines
while moving through the buffer.
Add support to it and add a setting to let the user adjust the number
of lines.

One downside of the implementation is that movement commands scroll
horizontally not just the needed movement but also scrolls making the
word position the left as possible. This is noticed when the scroll view
has horizontal scrolling and the user moves backward in word motions in
the words that are normally outside the view when the view is scrolled
maximum to the left.

The other downside is that insert mode doesn't take scroll off into
account. But the view position right after entering modal mode. I guess
this is good enough for now.

As a detail, a scrolloff greater than half of the number of lines in the
buffer keeps the current line always in the middle.

https://bugzilla.gnome.org/show_bug.cgi?id=742442